backend/requirements.txt
#lista de dependências - backend
flask
psycopg2-binary
flask-sqlalchemy
python-dotenv
